remote drag content component
