+++ client/src/components/main-content/SectionList.jsx
@@ -20,12 +20,12 @@ import { apiFetch } from "../../utils/apiFetch";
 import DeleteButton from "./Delete/DeleteButton";
 import GhostComponent from "./Add/GhostComponent";
 import SectionModal from "./Sections/SectionModal";
-import useRemoteDrags from "../../hooks/useRemoteDrags";
 import useRoomUsers from "../../hooks/rooms/useRoomUsers";
 import useDragHandlers from "../../hooks/useDragHandlers";
 import useSaveHandlers from "../../hooks/useSaveHandlers";
 import DroppableSection from "./Sections/DroppableSection";
 import CursorOverlay from "./SectionListComponents/CursorOverlay";
+import RemoteDragContent from "./SectionListComponents/RemoteDragContent";
 import useRoomCursors from "../../hooks/rooms/useRoomCursors";
 import useRoomEditing from "../../hooks/rooms/useRoomEditing";
 import NewSectionDropZone from "./Sections/NewSectionDropZone";
@@ -93,7 +93,6 @@ const SectionList = () => {
   });
 
   const cursors = useRoomCursors(roomId, userId);
-  const remoteDrags = useRemoteDrags(roomId, userId);
 
   useEffect(() => {
     if (!isAuthenticated || !username) return;
@@ -268,63 +267,6 @@ const SectionList = () => {
     return null;
   })();
 
-  const remoteDragContent = Object.values(remoteDrags).map((drag) => {
-    let content = null;
-    let isSection = false;
-    if (drag.type === DraggableComponentTypes.SECTION && sections[drag.id]) {
-      content = sections[drag.id].title;
-      isSection = true;
-    } else if (drag.type === DraggableComponentTypes.ITEM) {
-      for (const section of Object.values(sections)) {
-        const item = section.items.find((i) => i.id === drag.id); // find item in section
-        if (item) {
-          content = item.content;
-          break;
-        }
-      }
-    }
-    if (!content) return null;
-
-    return (
-      <div
-        key={drag.userId}
-        style={{
-          position: "fixed",
-          left: drag.x,
-          top: drag.y,
-          pointerEvents: "none",
-          zIndex: 9999,
-          opacity: 0.7,
-          color: drag.color,
-          fontWeight: "bold",
-          transform: "translate(-50%, -50%)",
-          transition: "left 0.05s, top 0.05s",
-        }}
-      >
-        <div
-          style={{
-            minWidth: 150,
-            padding: isSection ? "24px 32px" : "12px 24px",
-            background: isSection ? "var(--dark2)" : "#25242d",
-            borderRadius: isSection ? 12 : 8,
-            boxShadow: isSection
-              ? "0 2px 12px rgba(0,0,0,0.14)"
-              : "0 2px 8px rgba(0,0,0,0.12)",
-            opacity: 0.7,
-            fontWeight: isSection ? 600 : 500,
-            color: "white",
-            fontSize: isSection ? 20 : undefined,
-            border: `4px solid ${drag.color}`,
-            transform: "scale(1)",
-            transition: "transform 0.1s, opacity 0.1s",
-          }}
-        >
-          {content}
-        </div>
-      </div>
-    );
-  });
-
   return (
     <>
       <Container className="section-list-container">
@@ -367,7 +309,7 @@ const SectionList = () => {
               <AddButton />
             </div>
             <DragOverlay dropAnimation={null}>{dragOverlayContent}</DragOverlay>
-            <div>{remoteDragContent}</div>
+            <RemoteDragContent roomId={roomId} userId={userId} sections={sections} />
           </DndContext>
         </div>
       </Container>
